% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccm.R
\name{ccm}
\alias{ccm}
\title{Convergent Cross Map}
\usage{
ccm(X, Y, tau = 2, d = 4, rep_count = 25, N1 = 10, N2 = 100,
  silent = FALSE, top_count = NULL)
}
\arguments{
\item{X}{first time series or vector}

\item{Y}{second time series or vector of the same length as X}

\item{tau}{stepsize of shadow_manifold}

\item{d}{dimension of shadow_manifold}

\item{rep_count}{number of slices per stretchlength in the bootstrap}

\item{N1}{minimum of length of stretches}

\item{N2}{maximum of length of stratches}

\item{top_count}{number of points per slice to test in the bootstrap}
}
\value{
an ccm object with parameters and results
}
\description{
Run the convergent Cross Map algorithm
}
\examples{
# res <- ccm(X, Y)
# plot(ts(res$cor_x))
# lines(ts(res$cor_y))
}

