\name{data.expr}
\alias{data.expr}
\docType{data}
\title{
Sample gene expression and gender data
}
\description{
Simulated gene expression data
}
\usage{data(data.expr)}
\format{
A 100x40 (probe x sample) data matrix of gene expression data
}
\examples{
data(data.expr)
data(data.gender)
}
\keyword{datasets}
