\name{data.gender}
\alias{data.gender}
\docType{data}
\title{
Sample gene expression and gender data
}
\description{
Simulated gender data that corresponds with \code{data.expr}
}
\usage{data(data.gender)}
\format{
A 40 element vector containing simulated gender information as a factor (M/F), corresponding to columns of data.expr
}
\examples{
data(data.expr)
data(data.gender)
}
\keyword{datasets}
