% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict_y.R
\name{predict_y}
\alias{predict_y}
\title{Predict y}
\usage{
predict_y(sm.Y, wt)
}
\arguments{
\item{sm.Y}{a shadow manifold}

\item{wt}{a vector of weights on which to calculate a spatial average}
}
\value{
the predicted value of y (all coordinates)
}
\description{
Predicts the value of an observation in series Y given the X-weights and nearest neighbours
}
\examples{
#none
}
\seealso{
\code{\link{nn_weights}}
}
\keyword{predict}

