% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shadow_manifold.R
\name{shadow_manifold}
\alias{shadow_manifold}
\title{Make Shadow manifold}
\usage{
shadow_manifold(v, tau = 1, E = 3)
}
\arguments{
\item{v}{a a vector containing the observed data}

\item{tau}{the step size}

\item{E}{the dimension}
}
\value{
array - the multi-dimensional vector (v(t), v(t-tau), ... v(t-(E-1)*tau))
}
\description{
Make Shadow manifold from observed data v with stepsize tau and dimension E.
}
\examples{
#none
}
\seealso{
\code{\link{ccm}}
}
\keyword{shadow}

