% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccm.R
\name{ccm}
\alias{ccm}
\title{Convergent Cross Map}
\usage{
ccm(X, Y, tau = 2, d = 4, rep_count = 25, N1 = 10, N2 = 150)
}
\arguments{
\item{X}{first time series or vector}

\item{Y}{second time series or vector of the same length as X}

\item{tau}{stepsize of shadow_manifold}

\item{d}{dimension of shadow_manifold}

\item{rep_count}{number of elements in the bootstrap}

\item{N1}{minimum of length of stretches}

\item{N2}{maximum of length of stratches}
}
\value{
the two correlation vectors in a list
}
\description{
Run the convergent Cross Map algorithm
}
\examples{
# res <- ccm(X, Y)
# plot(ts(res$cor_x))
# lines(ts(res$cor_y))
}
\seealso{
\code{\link{get_nn}}
}
\keyword{neighbours}

