% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nearest_neighbours.R
\name{nearest_neighbours}
\alias{nearest_neighbours}
\title{Get nearest neighbours}
\usage{
nearest_neighbours(m, dists)
}
\arguments{
\item{m}{a shadow manifold}

\item{dists}{a matrix of euclidian distances between all points on the manifold x}
}
\value{
vector of indices of the E+1 nearest neighbours
}
\description{
Finds the indicies of the nearest neighbours to a point on a manifold
}
\examples{
# none
}
\seealso{
\code{\link{get_nn}}
}
\keyword{neighbours}

