% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nn_weights.R
\name{nn_weights}
\alias{nn_weights}
\title{Calculate weights}
\usage{
nn_weights(m, dists, index, nn)
}
\arguments{
\item{m}{a matrix containing a shadow manifold // we don't need it here yet}

\item{dists}{a matrix of euclidian distances between all points on the manifold x}

\item{index}{containing the index of the point in m .}

\item{nn}{a vector of indicies of the nearest neighbours}
}
\value{
list of weights and indices
}
\description{
Calculate "spatial" weights for each of the (d+1) nearest neighbours
}
\examples{
#none
}
\seealso{
\code{\link{get_nn}}
}
\keyword{weights}

