% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shadow_manifold.R
\name{shadow_manifold}
\alias{shadow_manifold}
\title{Make Shadow manifold}
\usage{
shadow_manifold(v, tau = 1, E = 4)
}
\arguments{
\item{v}{a a vector containing the observed data}

\item{tau}{the step size}

\item{E}{the dimension}
}
\value{
the distance matrix of the shadow manifold
}
\description{
Make distance matrix for Shadow manifold of vector v
}
\examples{
#none
}
\seealso{
\code{\link{ccm}}
}
\keyword{shadow}

